refactor(shared): create @nexus/shared workspace package (#11)
Move canonical DB types (DBNode, DBEntity, DBEdge, ReviewWithNode, AuthUser) from apps/web/src/lib/types.ts into packages/shared/src/types.ts. The original file becomes a re-export shim so all 10 existing import sites continue to work unchanged.
