+++ packages/shared/src/types.ts
@@ -1,0 +1,48 @@
+// ============================================================================
+// Canonical database types — matches supabase_setup.sql
+// Shared across apps/web and apps/extension via @nexus/shared.
+// ============================================================================
+
+export interface DBNode {
+  id: string
+  user_id: string
+  url: string
+  title: string
+  summary: string
+  raw_text: string
+  created_at: string
+}
+
+export interface DBEntity {
+  id: string
+  name: string
+  type: string
+  user_id: string
+  node_id: string | null
+}
+
+export interface DBEdge {
+  id: string
+  source_id: string
+  target_id: string
+  relation_type: string
+  weight: number
+  user_id: string
+}
+
+export interface ReviewWithNode {
+  id: string
+  user_id: string
+  node_id: string
+  next_review_date: string
+  interval: number
+  ease_factor: number
+  last_reviewed_at: string | null
+  node: Pick<DBNode, 'id' | 'title' | 'summary' | 'url'>
+}
+
+/** Minimal user shape hydrated from Supabase auth into Zustand */
+export interface AuthUser {
+  id: string
+  email: string
+}
